np.int is deprecated in numpy>=1.20 - use np.int32 instead (#205)

+++ pretty_midi/instrument.py
@@ -186,7 +186,7 @@ class Instrument(object):
             return piano_roll
         piano_roll_integrated = np.zeros((128, times.shape[0]))
         # Convert to column indices
-        times = np.array(np.round(times*fs), dtype=np.int)
+        times = np.array(np.round(times*fs), dtype=np.int32)
         for n, (start, end) in enumerate(zip(times[:-1], times[1:])):
             if start < piano_roll.shape[1]:  # if start is >=, leave zeros
                 if start == end:
